Separated custom keyboard util functions to reusable util functions in a new util file
